Use test template
