+++ packages/dx-react-chart-demos/src/demo-sources/zoom-pan/scatter.partial.jsxt
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import * as React from 'react';<%&additionalImports%>
 import {
   ArgumentScale,
   ValueScale,
@@ -13,69 +13,62 @@ import {
   ScatterSeries,
   Tooltip,
   ZoomAndPan,
-} from '@devexpress/dx-react-chart-bootstrap4';
+  Title,
+  Legend,
+} from '@devexpress/dx-react-chart-<%&themeName%>';
+<%&imports%>
 import { format } from 'd3-format';
 
-import { dataGenerator } from '../../../demo-data/generator';
+import { lifeBirth } from '../../../demo-data/life-birth';
+
+const data = lifeBirth.map(({
+  birthRate, lifeExp, year, country,
+}) => ({
+  [`birthRate_${year}`]: birthRate,
+  [`lifeExp_${year}`]: lifeExp,
+  year,
+  country,
+}));
 
 const adjustDomain = ([start, end]) => [Math.floor(start), Math.ceil(end)];
 
 const formatTooltipValue = format('.2f');
-const tooltipContentArgStyle = {
-  paddingBottom: 0,
-};
-const tooltipContentValStyle = {
-  paddingTop: 0,
-};
-const TooltipContent = ({ targetItem, data, ...restProps }) => {
+
+<%&body%>
+
+const TooltipContent = ({ targetItem, ...restProps }) => {
   const item = data[targetItem.point];
-  const arg = item[targetItem.series === 'Series 1' ? 'arg1' : 'arg2'];
-  const val = item[targetItem.series === 'Series 1' ? 'val1' : 'val2'];
+  const title = `${item.country} ${item.year}`;
+  const arg = item[`lifeExp_${targetItem.series}`];
+  const val = item[`birthRate_${targetItem.series}`];
   return (
-    <div>
-      <div>
-        <Tooltip.Content
-          {...restProps}
-          style={tooltipContentArgStyle}
-          text={`X: ${formatTooltipValue(arg)}`}
-        />
-      </div>
-      <div>
-        <Tooltip.Content
-          {...restProps}
-          style={tooltipContentValStyle}
-          text={`Y: ${formatTooltipValue(val)}`}
-        />
-      </div>
-    </div>
+    <TooltipContentBase
+      title={title}
+      arg={`Life Expectancy: ${formatTooltipValue(arg)}`}
+      val={`Birth Rate: ${formatTooltipValue(val)}`}
+      {...restProps}
+    />
   );
 };
-
-const ResetButton = props => (
-  <button type="button" className="btn btn-outline-primary m-2" {...props}>Reset</button>
-);
 
 export default class Demo extends React.PureComponent {
   constructor(props) {
     super(props);
 
-    const data = dataGenerator(200);
     this.state = {
       data,
       viewport: null,
     };
     this.changeViewport = viewport => this.setState({ viewport });
     this.resetViewport = () => this.setState({ viewport: null });
-
-    this.TooltipContent = allProps => <TooltipContent {...allProps} data={data} />;
   }
 
   render() {
     const { data, viewport } = this.state;
     return (
-      <div>
+      <React.Fragment>
         <ResetButton onClick={this.resetViewport} />
-        <div className="card">
+        <<%&wrapperTag%><%&wrapperAttributes%>>
           <Chart data={data}>
             <ArgumentScale modifyDomain={adjustDomain} />
             <ValueScale modifyDomain={adjustDomain} />
@@ -83,29 +76,31 @@ export default class Demo extends React.PureComponent {
             <ValueAxis />
 
             <ScatterSeries
-              name="Series 1"
-              valueField="val1"
-              argumentField="arg1"
+              name="1970"
+              valueField="birthRate_1970"
+              argumentField="lifeExp_1970"
             />
             <ScatterSeries
-              name="Series 2"
-              valueField="val2"
-              argumentField="arg2"
+              name="2010"
+              valueField="birthRate_2010"
+              argumentField="lifeExp_2010"
             />
 
             <EventTracker />
             <HoverState />
-            <Tooltip contentComponent={this.TooltipContent} />
+            <Tooltip contentComponent={TooltipContent} />
 
             <ZoomAndPan
               viewport={viewport}
               onViewportChange={this.changeViewport}
               interactionWithValues="both"
             />
+            <Title text="Life Expectancy vs. Birth Rates" />
+            <Legend />
             <Animation />
           </Chart>
-        </div>
-      </div>
+        </<%&wrapperTag%>>
+      </React.Fragment>
     );
   }
 }
